+++ R/twdtw.R
@@ -13,7 +13,7 @@
 ###############################################################
 
 .twdtw = function(x, y, weight.fun, dist.method, step.matrix, 
-                  n, span, min.length, theta, keep){
+                  n, span, min.length, keep){
   labels = as.character(labels(y))
   names(labels) = labels
   timeseries = x[[1]]